config/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-cors-headers==4.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.2
ffmpeg==1.4
gunicorn==21.2.0
h11==0.14.0
idna==3.4
inflection==0.5.1
lxml==4.9.3
mojito2==0.1.6
numpy==1.25.1
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pycparser==2.21
pycryptodome==3.18.0
pydub==0.25.1
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.11.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
SpeechRecognition==3.10.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==2.0.4
webdriver-manager==4.0.0
websockets==11.0.3
wsproto==1.2.0
